Rename ChannelList to GroupChannelList in the GroupChannelList module
